feature(server logic): add receiver logic to already existing initializer and sender logic
Our app is currently emulating the behaviour of a chat app server, by providing both sender and receiver logic. Furthermore, we are encapsulating in the receiver logic the mutation process of received messages into a list of client-with-message references and, also, we've managed to keep the connection's threads in an idle state while not being used.